modificacion migracion y vista welcome

+++ database/migrations/2024_05_20_051539_create_servicios_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->string('título');
             $table->text('descripción');
             $table->string('categoría');
-            $table->decimal('precio', 10, 2);
+            $table->integer('precio');
             $table->string('ubicación');
             $table->timestamp('fecha_publicación')->nullable();
             $table->boolean('activo')->default(true);
